refactor(restaurant-detail): ReviewItem 파일 이름 변경
